I been tried to add a button at the top of the main page, but it seems like since it need to use navigator as well as stack.navigator, so there is an conflict, I will figure out how to add the login button next time
