+++ HackRPIEventApp2023/App.js
@@ -1,18 +1,17 @@
-import React, { useState } from "react";
+import React from "react";
 import { StyleSheet, View, Text, Button } from "react-native";
 import { NavigationContainer } from "@react-navigation/native";
 import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
-import { createStackNavigator } from "@react-navigation/stack";
 import { Feather } from "@expo/vector-icons";
 import { StatusBar } from "expo-status-bar";
 import Calander from "./Box/Calander";
 import Food from "./information/Food";
 import HackerQue from "./HackerQue/QueEntry.js";
 import { globalStyles } from "./styles";
+import LoginButton from "./Login/LoginButton";
 import Login from "./Login/Login";
 
 const Tab = createBottomTabNavigator();
-const Stack = createStackNavigator();
 
 function InfoScreen() {
   return (
@@ -39,67 +38,52 @@ function QueueScreen() {
   );
 }
 
-const LoginScreen = ({ navigation }) => {
+const CustomHeader = ({ navigation }) => {
   return (
-    <View style={styles.container}>
-      <Button title="Login" onPress={() => navigation.navigate("Login")} />
+    <View style={{ flexDirection: 'row', justifyContent: 'space-between', paddingHorizontal: 20 }}>
+      <LoginButton navigation={navigation} />
     </View>
   );
 };
 
 export default function App() {
-  const [loggedIn, setLoggedIn] = useState(false);
-
   return (
     <NavigationContainer>
-      <Stack.Navigator>
-        {loggedIn ? (
-          <>
-            <Stack.Screen name="Home" component={HomeScreen} />
-            <Stack.Screen name="Info" component={InfoScreen} />
-            <Stack.Screen name="Queue" component={QueueScreen} />
-          </>
-        ) : (
-          <Stack.Screen name="Login" component={Login} />
-        )}
-      </Stack.Navigator>
-
-      {loggedIn && (
-        <Tab.Navigator
-          screenOptions={({ route }) => ({
-            tabBarIcon: ({ focused, color, size }) => {
-              let iconName;
-              if (route.name === "Info") {
-                iconName = "info";
-              } else if (route.name === "Home") {
-                iconName = "home";
-              } else if (route.name === "Queue") {
-                iconName = "user";
-              }
-              return (
-                <Feather
-                  name={iconName}
-                  size={size}
-                  color={focused ? "red" : "white"}
-                />
-              );
-            },
-            tabBarLabelStyle: {
-              fontSize: 12,
-            },
-            tabBarStyle: {
-              backgroundColor: globalStyles.primary,
-              borderTopWidth: 0,
-            },
-            tabBarActiveTintColor: globalStyles.accent,
-            tabBarInactiveTintColor: "white",
-          })}
-        >
-          <Tab.Screen name="Info" component={InfoScreen} />
-          <Tab.Screen name="Home" component={HomeScreen} />
-          <Tab.Screen name="Queue" component={QueueScreen} />
-        </Tab.Navigator>
-      )}
+      <Tab.Navigator
+        screenOptions={({ route }) => ({
+          header: (props) => <CustomHeader {...props} />,
+          tabBarIcon: ({ focused, color, size }) => {
+            let iconName;
+            if (route.name === "Info") {
+              iconName = "info";
+            } else if (route.name === "Home") {
+              iconName = "home";
+            } else if (route.name === "Queue") {
+              iconName = "user";
+            }
+            return (
+              <Feather
+                name={iconName}
+                size={size}
+                color={focused ? "red" : "white"}
+              />
+            );
+          },
+          tabBarLabelStyle: {
+            fontSize: 12,
+          },
+          tabBarStyle: {
+            backgroundColor: globalStyles.primary,
+            borderTopWidth: 0,
+          },
+          tabBarActiveTintColor: globalStyles.accent,
+          tabBarInactiveTintColor: "white",
+        })}
+      >
+        <Tab.Screen name="Info" component={InfoScreen} />
+        <Tab.Screen name="Home" component={HomeScreen} />
+        <Tab.Screen name="Queue" component={QueueScreen} />
+      </Tab.Navigator>
     </NavigationContainer>
   );
 }
